requirements.txt
dash==2.7.0
haversine== 2.8.0
pillow==9.4.0
streamlit_folium==0.11.1
textwrap==3 0.9.2
folium==0.14.0
matplotlib==3.7.1
numpy==1.24.3
pandas==1.5.3
pillow==9.4.0
plotly==5.9.0
plotly.express==0.4.1
re2==2022.04.01
streamlit==1.23.1
matplotlib-inline==0.1.6
